Rely on FakeStorage more
