+++ nexus/db-queries/src/db/datastore/auth.rs
@@ -79,95 +79,3 @@ impl Storage for super::DataStore {
         }
     }
 }
-
-#[cfg(test)]
-mod test {
-    use crate::db::DataStore;
-    use nexus_auth::authn;
-    use nexus_auth::authz::Action;
-    use nexus_auth::authz::AnyActor;
-    use nexus_auth::authz::Authz;
-    use nexus_auth::authz::Context;
-    use nexus_auth::authz::RoleSet;
-    use nexus_test_utils::db::test_setup_database;
-    use omicron_test_utils::dev;
-    use std::sync::Arc;
-
-    fn authz_context_for_actor(
-        log: &slog::Logger,
-        authn: authn::Context,
-        datastore: Arc<DataStore>,
-    ) -> Context {
-        let authz = Authz::new(log);
-        Context::new(Arc::new(authn), Arc::new(authz), datastore)
-    }
-
-    #[tokio::test]
-    async fn test_unregistered_resource() {
-        let logctx = dev::test_setup_log("test_unregistered_resource");
-        let mut db = test_setup_database(&logctx.log).await;
-        let (opctx, datastore) =
-            crate::db::datastore::test_utils::datastore_test(&logctx, &db)
-                .await;
-
-        // Define a resource that we "forget" to register with Oso.
-        use crate::context::OpContext;
-        use nexus_auth::authz::AuthorizedResource;
-        use omicron_common::api::external::Error;
-        use oso::PolarClass;
-        #[derive(Clone, PolarClass)]
-        struct UnregisteredResource;
-        impl AuthorizedResource for UnregisteredResource {
-            fn load_roles<'a, 'b, 'c, 'd, 'e>(
-                &'a self,
-                _: &'b OpContext,
-                _: &'c authn::Context,
-                _: &'d mut RoleSet,
-            ) -> futures::future::BoxFuture<'e, Result<(), Error>>
-            where
-                'a: 'e,
-                'b: 'e,
-                'c: 'e,
-                'd: 'e,
-            {
-                // authorize() shouldn't get far enough to call this.
-                unimplemented!();
-            }
-
-            fn on_unauthorized(
-                &self,
-                _: &Authz,
-                _: Error,
-                _: AnyActor,
-                _: Action,
-            ) -> Error {
-                // authorize() shouldn't get far enough to call this.
-                unimplemented!();
-            }
-
-            fn polar_class(&self) -> oso::Class {
-                Self::get_polar_class()
-            }
-        }
-
-        // Make sure an authz check with this resource fails with a clear
-        // message.
-        let unregistered_resource = UnregisteredResource {};
-        let authz_privileged = authz_context_for_actor(
-            &logctx.log,
-            authn::Context::privileged_test_user(),
-            Arc::clone(&datastore),
-        );
-        let error = authz_privileged
-            .authorize(&opctx, Action::Read, unregistered_resource)
-            .await;
-        println!("{:?}", error);
-        assert!(matches!(error, Err(Error::InternalError {
-            internal_message
-        }) if internal_message == "attempted authz check \
-            on unregistered resource: \"UnregisteredResource\""));
-
-        db.cleanup().await.unwrap();
-        logctx.cleanup_successful();
-    }
-}
